backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
httpx==0.27.2
pydantic==2.8.2
python-dotenv>=1.0
openai>=1.47
langgraph>=0.2.39
langgraph-checkpoint-sqlite>=2.0.9
langchain-core>=0.3
google-genai>=1.1.0
Pillow>=10.4
